+++ samples/ModelView/Model.h
@@ -70,11 +70,11 @@ public:
     virtual bool isValidIndex(const CEGUI::ModelIndex& model_index) const;
     virtual CEGUI::ModelIndex makeIndex(size_t child, const CEGUI::ModelIndex& parent_index);
     virtual CEGUI::ModelIndex getParentIndex(const CEGUI::ModelIndex& model_index);
-    virtual int getChildCount(const CEGUI::ModelIndex& model_index);
+    virtual size_t getChildCount(const CEGUI::ModelIndex& model_index);
     virtual CEGUI::String getData(const CEGUI::ModelIndex& model_index, CEGUI::ItemDataRole role = CEGUI::IDR_Text);
 
 private:
     std::vector<InventoryItem> d_inventoryItems;
 };
 
-#endif  // end of guard_Sample_Model_View_Model_h_
+#endif
